# Performance`cpu-max-all-1` +6%: using __bultin_expect within SeriesIteratorGetNext. Remove _seriesIteratorGetNext call