--- a/src/consts.h
+++ b/src/consts.h
@@ -9,6 +9,15 @@
 #include <sys/types.h>
 #include <stdbool.h>
 
+
+  #if defined(__GNUC__)
+#define likely(x)       __builtin_expect((x),1)
+#define unlikely(x)     __builtin_expect((x),0)
+  #elif _MSC_VER
+#define likely(x)       (x)
+#define unlikely(x)     (x)
+  #endif
+
 #define TRUE 1
 #define FALSE 0