Fixes #786 

### standalone improvements over `cpu-max-all-1`:
-  from 1457.90 to 1545.33 queries/sec **(~=6%)**
-  from p50 40.83ms to 38.59ms **(~=5.5%)**

To reproduce:
```
make clean build benchmark BENCHMARK=tsbs-scale100-cpu-max-all-1.yml
```

**master `cpu-max-all-1`:**
```
Run complete after 10000 queries with 64 workers (Overall query rate 1457.90 queries/sec):
RedisTimeSeries max of all CPU metrics, random    1 hosts, random 8h0m0s by 1h:
min:     0.93ms, med:    40.83ms, mean:    43.75ms, max:  104.28ms, stddev:    11.03ms, sum: 437.5sec, count: 10000
all queries                                                                   :
min:     0.93ms, med:    40.83ms, mean:    43.75ms, max:  104.28ms, stddev:    11.03ms, sum: 437.5sec, count: 10000
```

series.iterator.perf `cpu-max-all-1`:
```
Run complete after 10000 queries with 64 workers (Overall query rate 1545.33 queries/sec):
RedisTimeSeries max of all CPU metrics, random    1 hosts, random 8h0m0s by 1h:
min:     1.62ms, med:    38.59ms, mean:    41.28ms, max:   80.97ms, stddev:    10.20ms, sum: 412.8sec, count: 10000
all queries                                                                   :
min:     1.62ms, med:    38.59ms, mean:    41.28ms, max:   80.97ms, stddev:    10.20ms, sum: 412.8sec, count: 10000
```